requirements.txt
asgiref==3.8.1
defusedxml==0.7.1
diff-match-patch==20241021
Django==5.1.4
django-import-export==4.3.3
django-jazzmin==3.0.1
et_xmlfile==2.0.0
gunicorn==23.0.0
numpy==2.2.0
odfpy==1.4.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
tablib==3.7.0
tabulate==0.9.0
tzdata==2024.2
whitenoise==6.8.2
xlrd==2.0.1
xlwt==1.3.0
